# public/openapi.yaml
openapi: 3.1.0
info:
  title: DKIM Archive API
  description: |-
    This is the API specification for the DKIM Archive API, which follows the OpenAPI 3.0.3 specification.
  version: '0.1'
  license:
    name: MIT
    url: https://opensource.org/license/MIT
servers:
  - url: https://archive.prove.email/api
  - url: /api
    description: Development
paths:
...
  /dsp:
    post:
      tags:
        - domain-selector-pair
      summary: Add a domain/selector pair
      description: >
        Add a domain/selector pair.
        The server will try to fetch the DKIM key for the given domain and selector.
        If the key is found, the domain/selector pair will be added to the archive along with the key.
        The response indicates whether the domain/selector pair and the key were added to the archive, via the `added` field.
        The endpoint is rate limited to 1200 requests per 10 minutes.
      operationId: addDsp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainSelectorPair'
      responses:
        '201':
          description: Created - A new DKIM record was added to the archive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddDspResponse'
        '200':
          description: OK - Domain/selector pair processed but no new record added (already existed or key not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddDspResponse'
        '400':
          description: Invalid input
        '429':
          description: Rate limit exceeded
        '500':
          description: Unexpected error
...
components:
  schemas:
...
    DomainSelectorPair:
      type: object
      required:
        - domain
        - selector
      properties:
        domain:
          type: string
          examples:
            - example.com
        selector:
          type: string
          examples:
            - s1
...
    AddDspResponse:
      type: object
      required:
        - message
        - addResult
      properties:
        message:
          $ref: '#/components/schemas/DomainSelectorPair'
        addResult:
          type: object
          required:
            - already_in_db
            - added
          properties:
            already_in_db:
              type: boolean
              description: Whether the domain/selector pair was already in the database
            added:
              type: boolean
              description: Whether a new DKIM record was added to the archive